Fix filter placeholders for ui-select filter (frontend implementation)
