+++ app/code/Magento/Catalog/view/adminhtml/web/js/components/product-ui-select.js
@@ -3,6 +3,9 @@
  * See COPYING.txt for license details.
  */
 
+/**
+ * @deprecated see Magento/Ui/view/base/web/js/grid/filters/elements/ui-select.js
+ */
 define([
     'Magento_Ui/js/form/element/ui-select',
     'jquery',
