+++ app/code/Magento/MediaGalleryCatalogUi/view/adminhtml/ui_component/standalone_media_gallery_listing.xml
@@ -13,8 +13,7 @@
                     name="product_id"
                     provider="${ $.parentName }"
                     sortOrder="110"
-                    class="Magento\MediaGalleryCatalogUi\Ui\Component\Listing\Filters\UsedInProducts"
-                    component="Magento_Catalog/js/components/product-ui-select"
+                    component="Magento_Ui/js/grid/filters/elements/ui-select"
                     template="ui/grid/filters/elements/ui-select">
                 <argument name="data" xsi:type="array">
                     <item name="config" xsi:type="array">
