+++ app/code/Magento/MediaGalleryCmsUi/view/adminhtml/ui_component/media_gallery_listing.xml
@@ -14,7 +14,7 @@
                     provider="${ $.parentName }"
                     sortOrder="120"
                     class="Magento\MediaGalleryCmsUi\Ui\Component\Listing\Filters\UsedInPages"
-                    component="Magento_Ui/js/form/element/ui-select"
+                    component="Magento_Ui/js/grid/filters/elements/ui-select"
                     template="ui/grid/filters/elements/ui-select">
                 <argument name="data" xsi:type="array">
                     <item name="config" xsi:type="array">
@@ -27,6 +27,7 @@
                         <item name="emptyOptionsHtml" xsi:type="string" translate="true">Start typing to find pages</item>
                         <item name="filterRateLimit" xsi:type="string" translate="true">1000</item>
                         <item name="filterRateLimitMethod" xsi:type="string">notifyWhenChangesStop</item>
+                        <item name="validationUrl" xsi:type="url" path="media_gallery_cms/page/getSelected"/>
                     </item>
                 </argument>
                 <settings>
@@ -39,7 +40,7 @@
                     provider="${ $.parentName }"
                     sortOrder="130"
                     class="Magento\MediaGalleryCmsUi\Ui\Component\Listing\Filters\UsedInBlocks"
-                    component="Magento_Ui/js/form/element/ui-select"
+                    component="Magento_Ui/js/grid/filters/elements/ui-select"
                     template="ui/grid/filters/elements/ui-select">
                 <argument name="data" xsi:type="array">
                     <item name="config" xsi:type="array">
@@ -52,6 +53,7 @@
                         <item name="emptyOptionsHtml" xsi:type="string" translate="true">Start typing to find blocks</item>
                         <item name="filterRateLimit" xsi:type="string" translate="true">1000</item>
                         <item name="filterRateLimitMethod" xsi:type="string">notifyWhenChangesStop</item>
+                        <item name="validationUrl" xsi:type="url" path="media_gallery_cms/block/getSelected"/>
                     </item>
                 </argument>
                 <settings>
