+++ app/code/Magento/MediaGalleryUi/Ui/Component/Listing/Filters/Asset.php
@@ -16,9 +16,6 @@ use Magento\MediaContentApi\Api\GetContentByAssetIdsInterface;
 use Magento\Ui\Component\Filters\FilterModifier;
 use Magento\Ui\Component\Filters\Type\Select;
 use Magento\MediaGalleryApi\Api\GetAssetsByIdsInterface;
-use Magento\Cms\Helper\Wysiwyg\Images;
-use Magento\Cms\Model\Wysiwyg\Images\Storage;
-use Magento\Ui\Api\BookmarkManagementInterface;
 
 /**
  * Asset filter
@@ -36,21 +33,6 @@ class Asset extends Select
     private $getAssetsByIds;
 
     /**
-     * @var Images
-     */
-    private $images;
-
-    /**
-     * @var Storage
-     */
-    private $storage;
-
-    /**
-     * @var BookmarkManagementInterface
-     */
-    private $bookmarkManagement;
-
-    /**
      * Constructor
      *
      * @param ContextInterface $context
@@ -58,11 +40,6 @@ class Asset extends Select
      * @param FilterBuilder $filterBuilder
      * @param FilterModifier $filterModifier
      * @param OptionSourceInterface $optionsProvider
-     * @param GetContentByAssetIdsInterface $getContentIdentities
-     * @param GetAssetsByIdsInterface $getAssetsByIds
-     * @param BookmarkManagementInterface $bookmarkManagement
-     * @param Images $images
-     * @param Storage $storage
      * @param array $components
      * @param array $data
      * @SuppressWarnings(PHPMD.ExcessiveParameterList)
@@ -74,10 +51,6 @@ class Asset extends Select
         FilterModifier $filterModifier,
         OptionSourceInterface $optionsProvider = null,
         GetContentByAssetIdsInterface $getContentIdentities,
-        GetAssetsByIdsInterface $getAssetsByIds,
-        BookmarkManagementInterface $bookmarkManagement,
-        Images $images,
-        Storage $storage,
         array $components = [],
         array $data = []
     ) {
@@ -93,89 +66,6 @@ class Asset extends Select
             $data
         );
         $this->getContentIdentities = $getContentIdentities;
-        $this->getAssetsByIds = $getAssetsByIds;
-        $this->bookmarkManagement = $bookmarkManagement;
-        $this->images = $images;
-        $this->storage = $storage;
-    }
-
-    /**
-     * Prepare component configuration
-     *
-     * @return void
-     */
-    public function prepare()
-    {
-        $options = [];
-        $assetIds = $this->getAssetIds();
-
-        if (empty($assetIds)) {
-            parent::prepare();
-            return;
-        }
-
-        $assets = $this->getAssetsByIds->execute($assetIds);
-
-        foreach ($assets as $asset) {
-            $assetPath = $this->storage->getThumbnailUrl($this->images->getStorageRoot() . $asset->getPath());
-            $options[] = [
-                'value' => (string) $asset->getId(),
-                'label' => $asset->getTitle(),
-                'src' => $assetPath
-            ];
-        }
-
-        $this->optionsProvider = $options;
-        parent::prepare();
-    }
-
-    /**
-     * Get asset ids from filterData or from bookmarks
-     */
-    private function getAssetIds(): array
-    {
-        $assetIds = [];
-
-        if (isset($this->filterData[$this->getName()])) {
-            $assetIds = $this->filterData[$this->getName()];
-
-            if (!is_array($assetIds)) {
-                $assetIds = $this->stringToArray($assetIds);
-            }
-
-            return $assetIds;
-        }
-
-        $bookmark = $this->bookmarkManagement->getByIdentifierNamespace(
-            'current',
-            $this->context->getNameSpace()
-        );
-
-        if ($bookmark === null) {
-            return $assetIds;
-        }
-
-        $applied = $bookmark->getConfig()['current']['filters']['applied'];
-
-        if (isset($applied[$this->getName()])) {
-            $assetIds = $applied[$this->getName()];
-        }
-
-        if (!is_array($assetIds)) {
-            $assetIds = $this->stringToArray($assetIds);
-        }
-
-        return $assetIds;
-    }
-
-    /**
-     * Converts string array from url-applier to array
-     *
-     * @param string $string
-     */
-    private function stringToArray(string $string): array
-    {
-        return explode(',', str_replace(['[', ']'], '', $string));
     }
 
     /**
@@ -191,7 +81,7 @@ class Asset extends Select
 
         $assetIds = $this->filterData[$this->getName()];
         if (!is_array($assetIds)) {
-            $assetIds = $this->stringToArray($assetIds);
+            $assetIds = explode(',', str_replace(['[', ']'], '', $assetIds));
         }
 
         $filter = $this->filterBuilder->setConditionType('in')
