+++ app/code/Magento/MediaGalleryUi/view/adminhtml/ui_component/cms_block_listing.xml
@@ -13,7 +13,7 @@
                 provider="${ $.parentName }"
                 sortOrder="10"
                 class="Magento\MediaGalleryUi\Ui\Component\Listing\Filters\Asset"
-                component="Magento_Ui/js/form/element/ui-select"
+                component="Magento_Ui/js/grid/filters/elements/ui-select"
                 template="Magento_MediaGalleryUi/grid/filters/elements/ui-select">
                 <argument name="data" xsi:type="array">
                     <item name="config" xsi:type="array">
@@ -24,6 +24,7 @@
                         <item name="searchOptions" xsi:type="boolean">true</item>
                         <item name="filterOptions" xsi:type="boolean">true</item>
                         <item name="searchUrl" xsi:type="url" path="media_gallery/asset/search" />
+                        <item name="validationUrl" xsi:type="url" path="media_gallery/asset/getSelected"/>
                         <item name="levelsVisibility" xsi:type="number">1</item>
                     </item>
                 </argument>
